scripts: fix data lost in summary
lmbench3 records all test results and system information in
the raw results, and then parses the raw results through
getsummary script.

Once the mount details in the system information contain the
string "size=64xxx", this will cause the data parsing in
getsummary to break and exit, and then we cannot see the
complete results in summary.out.

This patch makes the parsing logic more accurate and fixes
such parsing failures.

Signed-off-by: Yong He <alexyonghe@tencent.com>